Split choice setup components to separate file
